optimize MVVM architecture
