+++ app/src/main/java/com/netology/nmedia/repository/PostRepository.kt
@@ -7,5 +7,5 @@ interface PostRepository {
     fun get(): LiveData<Post>
     fun like()
     fun share()
-    fun formatCount(count: Int): String?
+
 }
