+++ app/src/main/java/com/netology/nmedia/repository/PostRepositoryInMemoryImpl.kt
@@ -4,8 +4,7 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import com.netology.MyApp.R
 import com.netology.nmedia.dto.Post
-import java.math.RoundingMode
-import java.text.DecimalFormat
+
 
 class PostRepositoryInMemoryImpl : PostRepository {
     private var post = Post(
@@ -36,28 +35,5 @@ class PostRepositoryInMemoryImpl : PostRepository {
         post = post.copy(shares = post.shares + 1)
         data.value = post
     }
-    override fun formatCount(count: Int):String? {
-        when (count) {
-            in 0..999 -> return count.toString()
-            in 1000..9999 -> {
-                val df = DecimalFormat("#.#")
-                df.roundingMode = RoundingMode.DOWN
-                val result = df.format(count / 1000.0)
-                return result.toString() + "K"
-            }
-            in 10000..999999 -> {
-                val df = DecimalFormat("#")
-                df.roundingMode = RoundingMode.DOWN
-                val result = df.format(count / 1000.0)
-                return result.toString() + "K"
-            }
-            in 1000000 .. 999999999 -> {
-                val df = DecimalFormat("#.#")
-                df.roundingMode = RoundingMode.DOWN
-                val result = df.format(count / 1000000.0)
-                return result.toString() + "M"
-            }
-        }
-        return null
-    }
+
 }
